internal/task0029/task0029.go
/*
https://leetcode.com/problems/merge-intervals/
*/

package task0029

import "sort"

func Solution(intervals [][]int) (ans [][]int) {
	sort.Slice(intervals, func(i, j int) bool {
		return intervals[i][0] < intervals[j][0]
	})
	for _, interval := range intervals {
		if len(ans) == 0 || ans[len(ans)-1][1] < interval[0] {
			ans = append(ans, interval)
		} else {
			ans[len(ans)-1][1] = max(ans[len(ans)-1][1], interval[1])
		}
	}
	return ans
}

func max(a, b int) int {
	if a > b {
		return a
	}
	return b
}
